emulta/requirements.txt
Django==1.3
South==0.7.3
django-email-confirmation==0.1.4
django-gmapi==1.0.1
django-oembed==0.1.3
gondor==1.0b1.post9
